backend/requirements.txt
# ─── Web Framework ──────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ─── Settings & Env ─────────────────────────────────────────────────
pydantic-settings==2.2.1
python-dotenv==1.0.1

# ─── Document Parsing ───────────────────────────────────────────────
PyPDF2==3.0.1
python-docx==1.1.2

# ─── Text Chunking ──────────────────────────────────────────────────
langchain-text-splitters==0.2.0

# ─── Embeddings ─────────────────────────────────────────────────────
sentence-transformers==2.7.0

# ─── Vector Store ───────────────────────────────────────────────────
chromadb==0.5.23

# ─── LLM Clients ────────────────────────────────────────────────────
groq==0.9.0
httpx==0.27.0

# ─── Utilities ──────────────────────────────────────────────────────
aiofiles==23.2.1

# ─── Testing ────────────────────────────────────────────────────────
pytest==8.2.2
httpx==0.27.0
